pyproject.toml
[project]
name = "pubmed-download-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.12.3",
    "biopython>=1.85",
    "ipykernel>=7.0.1",
    "pdbpp>=0.11.7",
    "pykernel>=0.1.6",
    "requests>=2.32.5",
]

